Apply UI enhancements
App is now triggered via a menubar icon. Sizing has been adjusted to better accomodate a large number of screen captures.
Share and open in browser icons have been added to UI, but are not yet functional. Delete button text was replaced with
an icon. Size and date modified metadata was added to each capture.